refactor: changed project folder structure
