+++ src/App.css
@@ -4,9 +4,4 @@
     background-color: #f3f2ef;
 }
 
-.app__body{
-    display: flex;
-    margin: 35px 250px;
-    max-width: 100%;
-}
