+++ src/App.js
@@ -1,19 +1,13 @@
 import React from 'react';
 import './App.css';
-import Feed from './Feed';
-import Header from './Header';
-import SideBar from './SideBar';
-import Widget from './Widgets';
+import Header from './components/Header/Header';
+import HomeScreen from './screens/HomeScreen/HomeScreen';
 
 function App() {
   return (
     <div className="app">
         <Header />
-          <div className="app__body">
-              <SideBar />
-              <Feed />
-              <Widget />
-          </div>
+        <HomeScreen />
     </div>
   );
 }
