+++ src/screens/HomeScreen/HomeScreen.css
@@ -1,0 +1,6 @@
+
+.home__body{
+    display: flex;
+    margin: 35px 250px;
+    max-width: 100%;
+}
